+++ drivers/soc/hobot/bpu/bpu_prio.h
@@ -26,7 +26,7 @@ struct bpu_prio_node {
 	 * continue set the left part.
 	 */
 	struct bpu_fc residue_bpu_fc;
-	uint32_t left_slice_num;
+	int32_t left_slice_num;
 };
 
 struct bpu_prio {
@@ -34,15 +34,15 @@ struct bpu_prio {
 	/* to find the high task and set to hw_tail */
 	struct tasklet_struct tasklet;
 	struct mutex mutex_lock;
-	int inited;
+	uint16_t inited;
 	/* array to store support prios */
 	struct bpu_prio_node *prios;
-	int level_num;
+	uint32_t level_num;
 };
 
-struct bpu_prio *bpu_prio_init(struct bpu_core *core, int levels);
+struct bpu_prio *bpu_prio_init(struct bpu_core *core, uint32_t levels);
 void bpu_prio_exit(struct bpu_prio *prio);
-int bpu_prio_in(struct bpu_prio *prio, struct bpu_fc *bpu_fc);
-int bpu_prio_trig_out(struct bpu_prio *prio);
+int32_t bpu_prio_in(struct bpu_prio *prio, struct bpu_fc *bpu_fc);
+int32_t bpu_prio_trig_out(struct bpu_prio *prio);
 
 #endif
